python/end-to-end-applications/rag-ingestion/requirements.txt
restate_sdk
hypercorn
minio
langchain
langchain_community
langchain-ollama
langchain-qdrant
pypdf